Fix compiler warnings

+++ src/RubberBandStretcher.cpp
@@ -42,7 +42,7 @@ RubberBandStretcher::RubberBandStretcher(size_t sampleRate,
     m_r3d
     ((options & OptionEngineFiner) ?
      new R3StretcherImpl(R3StretcherImpl::Parameters
-                         (sampleRate, channels, options),
+                         (double(sampleRate), channels, options),
                          initialTimeRatio, initialPitchScale)
      : nullptr)
 {
